pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Niko Gamsakhurdiya"]
readme = "README.md"
packages = [{include = "gendiff"}]

[tool.poetry.dependencies]
python = "^3.10"
mypy = "^1.0.0"
flake8 = "^5.0.4"
jsonschema = "^4.17.3"
coverage = "^7.1.0"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pyyaml = "^6.0"
faker = "^18.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gen_diff:main"